Add a pre-handler that triggers an input event when text is pasted into an xforms text input widget. Use this in the Preamble dialog.
If anybody wants to help add this functionality to the rest of the dialogs
then feel free. Apart from FormPreferences (which has it's own prehandler
already that would need to be modified) there are about 60 text input
widgets that all need the extra code:
	extern "C" int C_CutandPastePH(FL_OBJECT *, int, FL_Coord, FL_Coord,
				       int, void *);

	fl_set_object_prehandler(widget, C_CutandPastePH);

Not too hard, is it?

Angus


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@2613 a592a061-630c-0410-9148-cb99ea01b6c8