Stabilise the tooltips interface and use it successfully in the Bibtex, Citation and Texinfo dialogs.
Remove all the horrible #if FL_REVISION < 89 cruft from FormBase and
put it in a little wrapper class Tooltips. For now use the 0.88 code for
all versions of xforms as I can't see an easy way of changing the tooltip
on the fly with fl_set_object_helper.

To use, call setTooltipHandler for each widget, add a choice widget to
the dialog together with 4 lines of code to set/get the level of verbosity.
Finally, and two methods, getMinimalTooltip and getVerboseTooltip to return
the tooltip.

What could be easier? Please try this out.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@3496 a592a061-630c-0410-9148-cb99ea01b6c8