Fix default fonts in RH8.0. We use the style hint after all else fails in order to find a reasonable font.
Note that StyleSerif does not work. I blame RedHat 8.0 personally,
but cannot verify yet.

Note also that in a sane world, we could just use font.setFamily(blah);
QFontInfo(font).family() and set to that. But Qt is obviously not sane, so
we stay with the weird setComboxFont() code we have :(


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@5869 a592a061-630c-0410-9148-cb99ea01b6c8