make the formats/converters panels in the pref dialog set the correct parameters.
atm this does not work because it depends on the order in which these entries were added.

now the order is alphabetical and things are messed up.

the attached patch adds the nr of the format/converter to the listwidgetitem as a type and uses that to get to the right details


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@13864 a592a061-630c-0410-9148-cb99ea01b6c8