* guiapi.[Ch]: deleted.
Angus L. explanation: Lars introduced it years ago when he got exited by the idea 
of dll-importing the frontend library, but the idea never came to anything 
concrete. The file can always be resurrected later on; it's meant to be a C-
language wrapper to our C++ frontend library calls.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@15611 a592a061-630c-0410-9148-cb99ea01b6c8