Fix bug 3356  by Bernhard Roider:
The crash happens on alt+up and alt+down if the position of the cursor
in the moved paragraph is beyond the length of the target paragraph.

To avoid it the Position of the iterators for the target paragraph must
be corrected. I set it to 0 as i assume it is not really used.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@17613 a592a061-630c-0410-9148-cb99ea01b6c8