Fix bug 2466 by Stefan Schimanski:
If the cursor is currently in macro mode (e.g. the user is typing \fra, but
hasn't finished the command \frac) macros shouldn't be updated.
Another forced screen update is needed if the user leaves macro mode just with
the cursor right key (which doesn't trigger full screen updates itself).


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@17860 a592a061-630c-0410-9148-cb99ea01b6c8