Fix arabic character transformation by Mostafa Vahedi:
The old code in the method "Paragraph::transformChar" had a bug, when the prev_char is a COMPOSE character (such as FATHA) and the previous BASE character is a SPECIAL character (such as ALEF). In this case the current character (such as BAA) gets the medial form and gets connected to the previous SPECIAL character which is not correct. The correct way to handle this situation is that, similar to the next_char, the first previous non-COMPOSE character should be assigned to the prev_char if exists.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@18065 a592a061-630c-0410-9148-cb99ea01b6c8