Scons: new update_manifest target.
When you run 'scons update_manifest', it tells you which files are missing
and which files are not in the source tree. It also generates a 
scons_manifest.py.new file with all the missing files added to 
XXX_extra_files. It will *not* change other sections of existing
manifest.py



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@18121 a592a061-630c-0410-9148-cb99ea01b6c8