Fix some UI bugs related to the paragraph settings dialog. The default checkbox has been changed to a radio button, so that other alignments can be chosen without the user's having to uncheck "Default".
There was some disagreement about the interface here. See, for example,
this thread:
  http://www.mail-archive.com/lyx-devel@lists.lyx.org/msg121328.html
For now, this will do.

The interface here is not fully functional. At present, if the Default 
happens to be Justified, then checking Justified is simply equivalent to
checking Default. This is due to the behavior of Text::setParagraph(), and
it should be changed if this interface is retained.



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@18860 a592a061-630c-0410-9148-cb99ea01b6c8