Fix bug 3970: Include file crash
The problem is that a screen update needs a bibtex key for proper rendering of the Citation. This key requires the master document to load all child document, which trigger an additional screen update, even when we don't switch buffer. The problem is that we are already in the process to update the screen, hence the crash.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@18986 a592a061-630c-0410-9148-cb99ea01b6c8